extends default graphics object
