+++ src/phaser/gameObjects/index.ts
@@ -1,20 +1,11 @@
 import Phaser from "phaser"
 
+import "./Graphics"
 import "./Image"
 import Button from "./Button"
-import CustomGraphics from "./CustomGraphics"
 import FloatingActionButton from "./FloatingActionButton"
 
 Phaser.GameObjects.GameObjectFactory.register("button", Button)
-
-Phaser.GameObjects.GameObjectFactory.register(
-  "customGraphics",
-  function (
-    this: Phaser.GameObjects.GameObjectFactory,
-  ): Phaser.GameObjects.CustomGraphics {
-    return this.scene.add.existing(new CustomGraphics(this.scene))
-  },
-)
 
 Phaser.GameObjects.GameObjectFactory.register(
   "fab",
